Jmaster

+++ proj/app/controllers/comments_controller.rb
@@ -1,23 +1,32 @@
 class CommentsController < ApplicationController
 
+	def index
+		@comments = Comment.where(link_id: params[:link_id])
+ 		render json: @comments.to_json(methods: :commenter)
+	end
+
 	def create
-		@comment = Comment.new(comment_params)
-		@comment.user_id = current_user.id
-		if @comment.save
-			respond_to do |format|
-				format.html {redirect_to @comment.link, notice: "Comment sucessfully created"}
-				format.js
+			@comment = Comment.new(comment_params)
+			@comment.user_id = current_user.id
+		respond_to do |format|
+			if @comment.save
+				format.html
+				format.json{ render json: @comment.to_json(methods: :commenter) }
+			else
+				redirect_to link_path(@comment.link), alert: @comment.errors.full_messages.to_sentence
 			end
-		else
-			redirect_to link_path(comment.link)
 		end
 	end
 
 
 	def destroy
 		comment = Comment.find(params[:id])
-		comment.destroy
-		redirect_to comment.link
+		if comment.user == current_user || comment.link.user == current_user
+			comment.destroy
+		  redirect_to comment.link
+		else
+		  redirect_to comment.link, alert: "You are not authorized to delete this comment"
+		end
 	end
